feat: Verify cached Mini App files using SHA512 hash (#117)
* feat: Verify cached Mini App files using SHA512 hash

After downloading a Mini App, an SHA512 hash will be generated for all of the files. Next, a hash is generated for the sum of these hashes and then this final hash value is stored in the device KeyChain. If the files that are cached on the device are modified, then the hash value in the keychain will no longer match, so the Mini App will be deleted and re-downloaded.