[MINI-5710]-[iOS][Demo App]: Event listener- External WebView closed and MiniApp resumed count is not updated in offline mode
Issue: When External web-view is closed the MiniApp resumed count was not recorded when in flight mode.
Cause: The url parameter used to be sent in the close action call back was not available and webview was nil.
Fix: Make use of the parameter currentURL set during the initializing of the ViewController.

Link: https://jira.rakuten-it.com/jira/browse/MINI-5710