### [MINI-5697] [iOS SampleApp] - Unable to open any contact from settings. ---------------------------------------
Issue: Unable to open any contact from settings and edit the contact details.

Cause: onTapGesture handler was missing for the List view cells in new SwiftUI implementation

Fix: Added onTapGesture for the Contacts List cell, On tap the edit contact will be presented and can edit the contact and save it.

**Link:**
https://jira.rakuten-it.com/jira/browse/MINI-5697