### [NO-TICKET] Sample App issues while editing the contact.
---------------------------------------

Issue: When editing the contact, The empty values for Name,Id,EmailId are getting saved in to local property value, hence reflecting wrong values in UI.

Cause: Due to binding object the values are immediately updated to the view model. Or on cancel of editing the view model must be update with the stored value.

Fix: Adding onCancle() action to the ContactsFormView and implementing the ObservableObject on MAContact, Changed ContactsData sent to formView a observed object.

**Link:**
https://jira.rakuten-it.com/jira/browse/MINI-5697
Fix Demo:[https://rak.box.com/s/i8pxv9ek871kviqcjmuiv7firaiv50mi]