[MINI-5676][iOS] Prepare the SDK to handle MAUID v2.
MiniApp Unique ID generation in Link relies on partly on the phone number from the user.
As we are migrating to a version of Link with potentially no phone number, we need to enrich the MAUID callback from the MiniApp SDK with a new variable containing a new identifier that does not rely on phone number.

Changes: Adding the deprecated modifier to the method getUniqueId() method.

**Link:**
[MINI-5676](https://jira.rakuten-it.com/jira/browse/MINI-5676)