[MINI-6076] Primary and secondary color
**New Feature: **
In order to be able to develop UI components as a MiniApp, as a MiniApp developer I want to adapt the UI of my host app, I want to be able to call a getHostColors bridge method to retrieve primary and secondary color of the host. In IOS SDK provide an interface to the host app that takes this 2 colors as parameters and call it whenever JS bridge calls the getHostColors method.
- Implemented Code changes in SDK and Sample app.
- Updated change log and user guide with new feature implementation details.

**Link:**
[MINI-6076](https://jira.rakuten-it.com/jira/browse/MINI-6076)