[MINI-6120] Add deprecated annotations for the old architecture
Added deprecated annotation for the methods in MiniApp class which will be removed from SDK in next releases.
Note: Some of the methods in this class are called from the demoApp even with the new architecture changes hence must be looked into moving the logic to MiniAppViewHandler or use the apis from hander of new architecture code.

**Link:**
[MINI-6120](https://jira.rakuten-it.com/jira/browse/MINI-6120)